move coords_3d_to_2d to nilearn.plotting.displays._utils
